chore: remove JournalStorageProvider
